% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_matches.R
\name{as_dataframe_mp_matches}
\alias{as_dataframe_mp_matches}
\title{as_dataframe_mp_matches}
\usage{
as_dataframe_mp_matches(g)
}
\arguments{
\item{g}{An igraph object representing the graph.}
}
\value{
A data frame with columns 'ENSG', 'repression_matches', and 'activation_matches' containing the vertex names and their respective feature values.
}
\description{
This function converts the graph's vertex attributes 'feature_rep' and 'features_act' into a data frame.
}
