% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_preds}
\alias{benchmark_preds}
\title{Benchmark seed match predictions}
\usage{
benchmark_preds(res, preds_col = "mer7m8", padj = 0.15, l2fc = 0)
}
\arguments{
\item{res}{A \code{data.frame} of \code{DESeq2} results.}

\item{preds_col}{Column name containing seed match counts.}

\item{padj}{Adjusted p-value cutoff for gold standard.}

\item{l2fc}{Log2(fold change) cutoff for gold standard.}
}
\value{
Two figures:
\itemize{
\item A precision recall curve
\item A receiver operator curve
}
}
\description{
Benchmark seed match predictions
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
benchmark_preds(res)
\dontshow{\}) # examplesIf}
}
