% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{build_annotation_filter}
\alias{build_annotation_filter}
\title{Build an \link[AnnotationFilter:AnnotationFilter]{AnnotationFilter::AnnotationFilter} for selecting transcripts}
\usage{
build_annotation_filter(
  standard.chroms = TRUE,
  remove.na.rows = TRUE,
  protein.coding = TRUE,
  transcript.support = NULL,
  canonical = FALSE,
  gene_id = NULL,
  tx_id = NULL,
  symbol = NULL,
  entrez_id = NULL,
  add.filter = NULL
)
}
\arguments{
\item{standard.chroms}{Keep only standard chromosomes. Boolean.}

\item{remove.na.rows}{Remove transcripts with NA in the ID column. Boolean.}

\item{protein.coding}{Keep only protein coding genes. Boolean.}

\item{transcript.support}{Filter by transcript support level. Integer 1-4}

\item{canonical}{Keep only the ENSEMBL canonical transcript. Boolean.}

\item{gene_id}{Keep specific ENSEMBL ID(s). Input is a vector}

\item{tx_id}{Keep specific tx_id(s). Input is a vector}

\item{symbol}{Keep specific gene symbol(s). Input is a vector}

\item{entrez_id}{Keep specific ENTREZ ID(s). Input is a vector}

\item{add.filter}{Optional \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} to include}
}
\value{
An \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} object
}
\description{
This function helps create an \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} that will combine
multiple standard filters based on the flags.
}
\examples{
# Default: standard chromosomes, removes na tx_ids, and keep protein coding
afl = SeedMatchR::build_annotation_filter()

# Keep only transcripts with support level 1
afl = SeedMatchR::build_annotation_filter(transcript.support = 1)

}
