% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{check_bulges_nt}
\alias{check_bulges_nt}
\title{check_bulges_nt}
\usage{
check_bulges_nt(sequenceStringSet, nt = nt, return_allowed = TRUE)
}
\arguments{
\item{sequenceStringSet}{An RNAStringSet object containing the sequences to be checked.}

\item{nt}{An integer representing the length from the start of the sequence where bulges are not allowed.}

\item{return_allowed}{A boolean indicating whether to return the allowed sequences (TRUE) or the removed sequences (FALSE).}
}
\value{
An RNAStringSet object containing either the allowed or removed sequences based on the \code{return_allowed} parameter.
}
\description{
This function checks for sequences with bulges and filters them based on the presence of 'N' in the first \code{nt} positions.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
