% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_list_overlap.R
\name{check_gene_list_overlap}
\alias{check_gene_list_overlap}
\title{Check if input gene lists overlap}
\usage{
check_gene_list_overlap(gene.lists)
}
\arguments{
\item{gene.lists}{A list of gene lists.
example: list(c("gene1", "gene2"), c("gene1"))}
}
\value{
Warning if gene sets overlap
}
\description{
Check if input gene lists overlap
}
\examples{
# Overlap
check_gene_list_overlap(list(c("gene1", "gene2"), c("gene1")))

#No overlap
check_gene_list_overlap(list(c("gene1", "gene2"), c("gene3")))
}
