% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph}
\alias{create_graph}
\title{Create a graph of protein-protein interactions from stringDB}
\usage{
create_graph(string_ppi, res)
}
\arguments{
\item{string_ppi}{The protein-protein interaction network.}

\item{res}{The result_output data frame.}
}
\value{
A graph object with the feature attribute.
}
\description{
This function creates a graph object from the protein-protein interaction network and the result_output dataframe.
}
