% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_fc_ecdf.R
\name{.check_gene_list_overlap}
\alias{.check_gene_list_overlap}
\title{Check if input gene lists overlap}
\usage{
.check_gene_list_overlap(gene_lists)
}
\arguments{
\item{gene_lists}{A list of gene lists.
example: list(c("gene1", "gene2"), c("gene1"))}
}
\value{
Warning if gene sets overlap.
}
\description{
Check if input gene lists overlap
}
\examples{
# Overlap
SeedMatchR:::.check_gene_list_overlap(list(c("gene1", "gene2"), c("gene1")))

#No overlap
SeedMatchR::.check_gene_list_overlap(list(c("gene1", "gene2"), c("gene3")))
}
