% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{.create_alignment_df}
\alias{.create_alignment_df}
\title{Create a dataframe of alignment information}
\usage{
.create_alignment_df(
  sirna,
  match.string,
  feature.seq,
  aln.name,
  left.pad,
  right.pad
)
}
\arguments{
\item{sirna}{The siRNA string as a character vector}

\item{match.string}{The match characters as a character vector}

\item{feature.seq}{The feature sequence as a character vector}

\item{aln.name}{The column name to use for the DF}

\item{left.pad}{The amount of left spaces to add based on upper-level arguments}

\item{right.pad}{The amount of right spaces to add based on upper-level arguments}
}
\value{
A \code{data.frame} that represents the alignment of the siRNA and feature sequence
}
\description{
Create a dataframe of alignment information
}
