% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{.find_factorial_internal}
\alias{.find_factorial_internal}
\title{find_factorial_internal}
\usage{
.find_factorial_internal(num_U)
}
\arguments{
\item{num_U}{An integer representing the number of Uridines.}
}
\value{
The sum of combinations as an integer.
}
\description{
This function calculates the sum of combinations for a given number of Uridines.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
