% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compensatory_sequences.R
\name{.gc_content}
\alias{.gc_content}
\title{.gc_content}
\usage{
.gc_content(sequence)
}
\arguments{
\item{sequence}{A character string representing the DNA sequence.}
}
\value{
A numeric value representing the GC content percentage of the sequence.
}
\description{
Function to calculate the GC content of a given DNA sequence.
}
