% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{.get_all_combination_w_b}
\alias{.get_all_combination_w_b}
\title{get_all_combination_w_b}
\usage{
.get_all_combination_w_b(guide.seq, silent = FALSE)
}
\arguments{
\item{guide.seq}{The guide sequence}

\item{silent}{A boolean to suppress output}
}
\value{
A list of guide and seed sequences
}
\description{
This function generates all possible combinations of guide sequences with wobbles
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
