% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{.get_canonical_seed}
\alias{.get_canonical_seed}
\title{Get the target sequence for canonical seed definitions}
\usage{
.get_canonical_seed(
  guide.seq,
  seed.name = c("mer7m8", "Full", "18mer", "15mer", "mer8", "mer6", "mer7A1"),
  allow_wobbles = FALSE
)
}
\arguments{
\item{guide.seq}{The input RNA string as a character vector. Must be at least
8 characters long.}

\item{seed.name}{The name of the canonical seed definition to return.
Options: "mer7m8", "mer8", "mer6", "mer7A1", "Full", "18-mer", "15-mer"}

\item{allow_wobbles}{If true, all positions for a G:U wobble are considered.}
}
\value{
A \code{DNAString} of the canonical seed sequence that is the search term.
}
\description{
Get the target sequence for canonical seed definitions
}
\examples{
guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

canonical.seed.seq = SeedMatchR:::.get_canonical_seed(guide.seq,
seed.name = "mer7m8")
}
