% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedMatchR.R
\name{.get_counts_column}
\alias{.get_counts_column}
\title{Get match results as a column of counts appended to a differential expression results \code{data.frame}}
\usage{
.get_counts_column(
  res,
  target.seq,
  seqs,
  seed.name,
  col.name = NULL,
  shared_genes = TRUE,
  max.mismatch = 0,
  with.indels = F,
  fixed = F
)
}
\arguments{
\item{res}{Input differential expression results \code{data.frame}}

\item{target.seq}{Pattern to search for in \code{seqs}}

\item{seqs}{The \code{DNAStringSet} to be search for instances of the \code{target.seq}}

\item{seed.name}{The seed name to be reported in the \code{data.frame} column
called \code{seed}}

\item{col.name}{The optional name of the column of match counts. Will default
to \code{seed.name} if not set}

\item{shared_genes}{If true, the tx set is reduced to overlapping
features in both the sequences db and in the DE data.}

\item{max.mismatch}{Number of allowed mismatches or the total edit distance}

\item{with.indels}{If True, include indels}

\item{fixed}{Require that each sequence symbol matches when searching. Should be FALSE if using wobbles.}
}
\value{
A column of match counts for each gene in \code{res}
}
\description{
Get match results as a column of counts appended to a differential expression results \code{data.frame}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

get_example_data("sirna")

sirna.data = load_example_data("sirna")

res <- sirna.data$Schlegel_2022_Ttr_D1_30mkg

#Load test data
res <- Schlegel_2022_Ttr_D1_30mkg

# Filter results for SeedMatchR
res = filter_res(res, fdr.cutoff=1, fc.cutoff=0, rm.na.log2fc = TRUE)

seed.seq = get_seed(guide.seq = guide.seq, seed.name = "mer7m8")

anno.db = load_annotations("rnor7")

SeedMatchR:::.get_counts_column(res = res, target.seq = seed.seq$Target.Seq,
seqs = anno.db$seqs, seed.name = "mer7m8")
\dontshow{\}) # examplesIf}
}
