% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed.R
\name{.get_custom_sequence}
\alias{.get_custom_sequence}
\title{Get a custom target sequence based on start and stop position}
\usage{
.get_custom_sequence(guide.seq, start.pos, stop.pos, allow_wobbles = FALSE)
}
\arguments{
\item{guide.seq}{The input RNA string as a character vector}

\item{start.pos}{Start position index. 1-based index}

\item{stop.pos}{Stop position index.}

\item{allow_wobbles}{If true, all positions for a G:U wobble are considered.}
}
\value{
A \code{DNAString} of the customized search sequence
}
\description{
This function will take a guide sequence and subset it based on custom start
and stop positions. The sequence will then be converted to the
reverse-complement \code{DNAString} of the input \code{RNAString} sequence. The start
and stop positions are 1-based indexes.
}
\examples{
guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

custom.seq = SeedMatchR:::.get_custom_sequence(guide.seq, start.pos = 1,
stop.pos = 23)
}
