% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{.get_longest_utr}
\alias{.get_longest_utr}
\title{Get the longest transcript per gene}
\usage{
.get_longest_utr(utr.grlist, txdb, filter = NULL)
}
\arguments{
\item{utr.grlist}{A \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} of UTR exons grouped by \code{tx_id}}

\item{txdb}{An \link[ensembldb:EnsDb]{ensembldb::EnsDb} object}

\item{filter}{Optional \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} object}
}
\value{
A \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} filtered containing the longest 3' UTR
per gene in the \link[ensembldb:EnsDb]{ensembldb::EnsDb} object that pass filters in the optional
\link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList}
}
\description{
\code{.get_longest_utr} will extract the longest transcript per gene in a \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList}
using an \link[ensembldb:EnsDb]{ensembldb::EnsDb} object. The \code{\link[GenomicFeatures:transcriptLengths]{GenomicFeatures::transcriptLengths()}} function
is used to calculate the 3' UTR lengths for each transcript in the \link[ensembldb:EnsDb]{ensembldb::EnsDb} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
annodb <- SeedMatchR::load_species_anno_db("hg38")

features = GenomicFeatures::threeUTRsByTranscript(annodb$txdb)

longest.tx = SeedMatchR:::.get_longest_tx(features, annodb$txdb)
\dontshow{\}) # examplesIf}
}
