% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{.get_target_bulged_guides}
\alias{.get_target_bulged_guides}
\title{get_target_bulged_guides}
\usage{
.get_target_bulged_guides(guide.seq, guide_seq_length, bulges_lim = bulges_lim)
}
\arguments{
\item{guide.seq}{A character vector representing the guide sequence.}

\item{guide_seq_length}{Integer, length of guide strand sequence}

\item{bulges_lim}{An integer representing the maximum number of bulges allowed.}
}
\value{
An RNAStringSet object containing the sequences with bulges.
}
\description{
This function generates a set of RNA sequences with bulges inserted at various positions.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
