% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{.make_space}
\alias{.make_space}
\title{Create a vector of spaces for padding sequences for alignment}
\usage{
.make_space(num.spaces)
}
\arguments{
\item{num.spaces}{Number of spaces to create in vector.}
}
\value{
A character vector of spaces the size of \code{num.spaces}
}
\description{
Create a vector of spaces for padding sequences for alignment
}
