% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{.matchStrings}
\alias{.matchStrings}
\title{Match two strings of the same size and return a character vector to represent positions that match}
\usage{
.matchStrings(s1, s2)
}
\arguments{
\item{s1}{String 1 to match}

\item{s2}{String 2 to match}
}
\value{
A character vector with \code{|} in the index for every match between both strings.
}
\description{
Match two strings of the same size and return a character vector to represent positions that match
}
\examples{
SeedMatchR:::.matchStrings("AAAACAGTGTTCTTGCTCTATAA", "ATAACAGTCTTCTTGCTCTGTAA")
}
