% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_fc_ecdf.R
\name{.plot_ecdf}
\alias{.plot_ecdf}
\title{Use ggplot2 to plot the ecdf for log2 fold changes}
\usage{
.plot_ecdf(
  res,
  palette = SeedMatchR.palette,
  title,
  legend_labels,
  test_results,
  theme = SeedMatchR.theme,
  x_lims = c(-1, 1),
  output_filename = NULL,
  height = 5,
  width = 5,
  dpi = 300
)
}
\arguments{
\item{res}{DE results data.frame.}

\item{palette}{Palette to use for line colors.}

\item{title}{Plot tile.}

\item{legend_labels}{Labels for legend entries.}

\item{test_results}{Results from the ecdf_stat_test function.}

\item{theme}{The ggplot2 theme parameters.}

\item{x_lims}{The x-axis limits for plotting.}

\item{output_filename}{Optional output filename.}

\item{height}{Height of output image in inches.}

\item{width}{Width of output image in inches.}

\item{dpi}{DPI of output image.}
}
\value{
A ggplot of the log2(fold changes) ECDF.
}
\description{
Use ggplot2 to plot the ecdf for log2 fold changes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
