% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{.reduce_features}
\alias{.reduce_features}
\title{Reduce 3' UTRs to the gene level across all transcripts}
\usage{
.reduce_features(gr.list, txdb, filter)
}
\arguments{
\item{gr.list}{Input \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} file with features grouped by transcript}

\item{txdb}{An \link[ensembldb:EnsDb]{ensembldb::EnsDb} object}

\item{filter}{An \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} for selecting transcripts}
}
\value{
A gtf of unique feature intervals per gene
}
\description{
Reduce 3' UTRs to the gene level across all transcripts
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
anno.db = SeedMatchR::load_species_anno_db("hg38")

features = SeedMatchR::get_features_gtf(anno.db$txdb)

gtf = SeedMatchR::.reduce_features(features, anno.db$txdb)
\dontshow{\}) # examplesIf}
}
