% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{.sequence_with_tb}
\alias{.sequence_with_tb}
\title{.sequence_with_tb}
\usage{
.sequence_with_tb(bulge_start, sequence, guide_seq_length)
}
\arguments{
\item{bulge_start}{An integer representing the start position of the bulge.}

\item{sequence}{Input sequence}

\item{guide_seq_length}{Integer, length of guide strand sequence}
}
\value{
A character string representing the sequence with the bulge.
}
\description{
This nested function generates a sequence with a bulge inserted at the specified start position.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
