% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compensatory_sequences.R
\name{.update_sequence_at}
\alias{.update_sequence_at}
\title{.update_sequence_at}
\usage{
.update_sequence_at(
  sequences,
  seed,
  mismatch_allowed = mismatch_,
  limit_ = limit
)
}
\arguments{
\item{sequences}{A DNAStringSet object containing the sequences of the transcripts.}

\item{seed}{A character string representing the seed sequence.}

\item{mismatch_allowed}{An integer representing the maximum number of mismatches allowed in the seed match.}

\item{limit_}{An integer representing the limit for sequence update.}
}
\value{
A DNAStringSet object containing the updated sequences and a data frame with the start sites of the seed matches.
}
\description{
Function to update the sequence library by extracting subsequences around the seed match.
}
