% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{download_parse_file}
\alias{download_parse_file}
\title{Download and parse DESeq2 output from GSE184929}
\usage{
download_parse_file(download.path, output.path)
}
\arguments{
\item{download.path}{File url to be downloaded}

\item{output.path}{Filename used for saving downloaded file}
}
\value{
DESeq2 results as a data.frame.
}
\description{
Download and parse DESeq2 output from GSE184929
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_parse_file()
\dontshow{\}) # examplesIf}
}
