% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_res.R
\name{filter_res}
\alias{filter_res}
\title{Filter differential expression results}
\usage{
filter_res(res, fdr_cutoff = 1, fc_cutoff = 0, basemean_cutoff = 30)
}
\arguments{
\item{res}{The differential expression results as a data frame}

\item{fdr_cutoff}{The false discovery rate cutoff to use.}

\item{fc_cutoff}{The fold change cutoff to use. The absolute
value will be used as the cutoff and values greater-than-or-equal-to will
be kept.}

\item{basemean_cutoff}{The minimum baseMean expression cutoff}
}
\value{
A filtered results data frame
}
\description{
Filter differential expression results for use with search
and counting functions.

The filtering criteria are:

Filter out genes that are not expressed or counted at all:
baseMean == NA | pvalue == NA | log2FoldChange == NA

Filter out genes with extreme outliers: pvalue = NA and padj = NA

Filter out genes that have been excluded by independent filtering. padj = NA

Filter results by the \code{fdr_cutoff}

Filter the results by the \code{log2FoldChange}

Filter the results by the \code{baseMean}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Load test data
get_example_data("sirna")

sirna.data = load_example_data("sirna")

res <- sirna.data$Schlegel_2022_Ttr_D1_30mkg

# Filter differential expression results for SeedMatchR
res = filter_res(res, fdr_cutoff=1, fc_cutoff=0)
\dontshow{\}) # examplesIf}
}
