% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_search.R
\name{full_search}
\alias{full_search}
\title{Search input \code{Biostrings::DNAStringSet} for a set of defined sequences}
\usage{
full_search(
  guide.seq,
  seqs,
  group.name = "siRNA",
  indel.bool = FALSE,
  allow_wobbles = FALSE,
  seqs.to.search = c("Full", "18mer", "15mer", "mer8", "mer7m8", "mer7A1", "mer6"),
  res_format = c("counts", "iranges"),
  fixed = FALSE,
  nonseed.edit.distances = c(0:4),
  seed.edit.distances = c(0:1)
)
}
\arguments{
\item{guide.seq}{Input guide sequence.}

\item{seqs}{\code{Biostrings::DNAStringSet} of sequences to search.}

\item{group.name}{Group name to use for labeling.}

\item{indel.bool}{Boolean in long format (TRUE, FALSE) for whether to include indels.}

\item{allow_wobbles}{If true, all positions for a G:U wobble are considered.}

\item{seqs.to.search}{List of sequences to search. Options: "Full", "18mer", "15mer", "mer8", "mer7m8", "mer7A1", "mer6".}

\item{res_format}{Either "iranges" or "counts".}

\item{fixed}{Whether to use matching string symbols for search when ambiguous.}

\item{nonseed.edit.distances}{An integer value indicating allowed edit distance for non-seed searches.}

\item{seed.edit.distances}{An integer value indicating allowed edit distance for seed searches.}
}
\value{
A \code{data.frame} of sequence match counts.
}
\description{
This function will use \code{Biostrings::vcountpattern} to find matches to the
input siRNA sequence. This function will then provide a matrix that is
sequences x match type. The best match will be reported based on a pre-defined
list of rankings.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
print("TBD")
\dontshow{\}) # examplesIf}
}
