% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_e_value.R
\name{generate_e_value}
\alias{generate_e_value}
\title{generate_e_value}
\usage{
generate_e_value(
  result_output,
  random_seq_results = random_siRNA_results_overall
)
}
\arguments{
\item{result_output}{A dataframe of siRNA sequences and their respective number of matches}

\item{random_seq_results}{A dataframe of random siRNA sequences and their respective number of matches}
}
\value{
A dataframe with the e-value for each gene
}
\description{
This function calculates the e-value for each gene in the result_output dataframe. The e-value is calculated as the number of times the random siRNA sequences have a higher number of matches than the actual siRNA sequences.
}
