% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_possible_seqs.R
\name{generate_possible_seqs}
\alias{generate_possible_seqs}
\title{Generate all possible sequence permutations given an alphabet and length}
\usage{
generate_possible_seqs(n = 8, alph = c("A", "U", "G", "C"))
}
\arguments{
\item{n}{number of characters in the sequence}

\item{alph}{the alphabet}
}
\value{
A data.frame containg all possible sequences. Each columns is a
position in the sequence.
}
\description{
This function is used to generate all possible sequences given
an alphabet (\code{alph}) and length of sequence (\code{n}).
}
\examples{
seq.df = generate_possible_seqs(n=8)
}
