% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{get_3UTR_granges}
\alias{get_3UTR_granges}
\title{Extract \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} of 3' UTRs}
\usage{
get_3UTR_granges(
  txdb,
  filter = NULL,
  longest.utr = FALSE,
  reduce.features = FALSE
)
}
\arguments{
\item{txdb}{An \link[ensembldb:EnsDb]{ensembldb::EnsDb} object}

\item{filter}{An \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} for selecting transcripts}

\item{longest.utr}{Reduce annotations to the longest UTR per gene}

\item{reduce.features}{Reduce 3' UTRs to non-overlapping intervals across all
transcripts per gene.}
}
\value{
A \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} object representing the genomic
features of interest. Results are returned grouped by transcript, unless
\code{longest.utr} is TRUE, then results are grouped by gene.
}
\description{
\code{get_3UTR_granges} leverages the \code{\link[GenomicFeatures:transcriptsBy]{GenomicFeatures::threeUTRsByTranscript()}}
function to extract 3' UTR regions from a given \link[ensembldb:EnsDb]{ensembldb::EnsDb} object.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
annodb <- SeedMatchR::load_species_anno_db("hg38")

features <- SeedMatchR::get_3UTR_granges(annodb$txdb)
\dontshow{\}) # examplesIf}
}
