% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compensatory_sequences.R
\name{get_compensatory_sequences}
\alias{get_compensatory_sequences}
\title{get_compensatory_sequences}
\usage{
get_compensatory_sequences(
  guide.seq,
  seed,
  sequences,
  compensatory_region_length_min = 6,
  compensatory_region_length_max = NA,
  limit = 100,
  mismatch_ = 1,
  mismatch_compensatory_region_ = 0
)
}
\arguments{
\item{guide.seq}{A character string representing the guide sequence.}

\item{seed}{A character string representing the seed sequence.}

\item{sequences}{A DNAStringSet object containing the sequences of the transcripts.}

\item{compensatory_region_length_min}{An integer representing the minimum length of the compensatory region.}

\item{compensatory_region_length_max}{An integer representing the maximum length of the compensatory region.}

\item{limit}{An integer representing the limit for sequence update.}

\item{mismatch_}{An integer representing the maximum number of mismatches allowed in the seed match.}

\item{mismatch_compensatory_region_}{An integer representing the maximum number of mismatches allowed in the compensatory region.}
}
\value{
A data frame containing the compensatory regions.
}
\description{
Function to get 3' compensatory sequences for a given siRNA sequence.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
