% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Download example DESeq2 data from GEO}
\usage{
get_example_data(example.type)
}
\arguments{
\item{example.type}{Name of the example to load. Options: sirna, mirna}
}
\value{
None
}
\description{
This function will download data that can be used  for SeedMatchR.
Choosing 'sirna' will download 3 DESeq2 results files from GSE184929. Choosing
'mirna' will download the miRDB database as a tsv.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_example_data()
\dontshow{\}) # examplesIf}
}
