% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_matches.R
\name{get_network_matches}
\alias{get_network_matches}
\title{get_network_matches}
\usage{
get_network_matches(
  ENSG_Genelist,
  organism,
  aggregator = "mean",
  network_data = NULL
)
}
\arguments{
\item{ENSG_Genelist}{A character vector of gene identifiers.}

\item{organism}{A character string specifying the organism. Options are 'human' or 'mouse'. Default is 'human'.}

\item{aggregator}{A character string specifying the aggregation method. Options are 'mean' or 'degree_importance'. Default is 'mean'.}

\item{network_data}{A data frame containing the transcription network data. If NULL, the function will download the data from a predefined URL.The columns of the data frame should be 'source', 'target', 'edge_type', and 'evidence'.}
}
\value{
A data frame with columns 'ENSG', 'repression_matches', and 'activation_matches' containing the vertex names and their respective feature values.
}
\description{
This function calculates network matches for a given graph and gene list using a transcription network and an aggregation method.
}
