% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stringPPI_to_stringGGI.R
\name{get_stringPPI_to_stringGGI}
\alias{get_stringPPI_to_stringGGI}
\title{get_stringPPI_to_stringGGI}
\usage{
get_stringPPI_to_stringGGI(sp_number, p = 0.95, sp = "rat")
}
\arguments{
\item{sp_number}{The species number for the STRINGdb object}

\item{p}{The percentile threshold for the STRINGdb network}

\item{sp}{The species for the Biomart object}
}
\value{
A protein-protein interaction network with gene ids
}
\description{
This function retrieves the protein-protein interaction network from STRINGdb and maps the protein ids to gene ids using Biomart.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
