% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_hist.R
\name{kmer_hist}
\alias{kmer_hist}
\title{Plot a histogram of kmer frequency across genes}
\usage{
kmer_hist(
  seqs,
  target.seq,
  width = 7,
  step = 1,
  title = "K-mer gene frequencies"
)
}
\arguments{
\item{seqs}{Input \link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet}}

\item{target.seq}{Character vector of target seed sequence. Must be same size as width parameter.}

\item{width}{Width of the kmer to count.}

\item{step}{Step size for kmer window along the sequence.}

\item{title}{Figure title}
}
\value{
A list of:
*\verb{$plot}: A histogram of frequencies X number of genes with kmer
*\verb{$matrix}: Counts matrix
}
\description{
Plot a histogram of kmer frequency across genes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
annodb = load_annotations("hg38", canonical = T)

kmer_hist(annodb$seqs, "AAAAAAA")
\dontshow{\}) # examplesIf}
}
