% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{load_annotations}
\alias{load_annotations}
\title{Load species-specific annotations needed to run \code{\link[=SeedMatchR]{SeedMatchR()}}}
\usage{
load_annotations(
  reference.name = c("hg38", "hg38-old", "mm39", "mm10", "rnor6", "rnor7"),
  feature.type = c("3UTR", "5UTR", "exons", "cds"),
  standard.chroms = TRUE,
  remove.na.rows = TRUE,
  protein.coding = TRUE,
  transcript.support = NULL,
  canonical = TRUE,
  gene_id = NULL,
  tx_id = NULL,
  symbol = NULL,
  entrez_id = NULL,
  longest.utr = FALSE,
  reduce.features = FALSE,
  add.filter = NULL,
  min.feature.width = 8,
  return_gene_name = TRUE
)
}
\arguments{
\item{reference.name}{Reference build name. Options: hg38, rnor6, rnor7,
mm39, mm10}

\item{feature.type}{The transcript feature type to extract. Options: "3UTR", "5UTR", "exons", "cds"}

\item{standard.chroms}{Keep only standard chromosomes. Boolean.}

\item{remove.na.rows}{Remove transcripts with NA in the ID column. Boolean.}

\item{protein.coding}{Keep only protein coding genes. Boolean.}

\item{transcript.support}{Filter by transcript support level. Integer 1-4}

\item{canonical}{Keep only the ENSEMBL canonical transcript. Boolean.}

\item{gene_id}{Keep specific ENSEMBL ID(s). Input is a vector}

\item{tx_id}{Keep specific tx_id(s). Input is a vector}

\item{symbol}{Keep specific gene symbol(s). Input is a vector}

\item{entrez_id}{Keep specific ENTREZ ID(s). Input is a vector}

\item{longest.utr}{Reduce annotations to the longest UTR per gene}

\item{reduce.features}{Reduce 3' UTRs to non-overlapping intervals across all
transcripts per gene.}

\item{add.filter}{Optional \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} to include}

\item{min.feature.width}{The minimum length for features in nucleotides.}

\item{return_gene_name}{If true, the gene_id will be used to name the sequences.}
}
\value{
The following are returned in the object:
\itemize{
\item \verb{$gtf}: A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing the transcript information
\item \verb{$txdb}: A \link[ensembldb:EnsDb]{ensembldb::EnsDb} object of transcript information
\item \verb{$dna}: The \link[rtracklayer:TwoBitFile-class]{rtracklayer::TwoBitFile} DNA sequence
\item \verb{$features}: A \link[GenomicRanges:GRangesList-class]{GenomicRanges::GRangesList} object representing the features of interest
\item \verb{$seqs}: A \link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet} for the features in \verb{$features}
\item \verb{$filter}: A \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} used for filtering the \link[ensembldb:EnsDb]{ensembldb::EnsDb} object
\item \verb{$tx.names}: A list of transcript names in \verb{$seqs}
}
}
\description{
\code{load_annotations} uses \code{\link[AnnotationHub:AnnotationHub-class]{AnnotationHub::AnnotationHub()}} to load
species-specific annotations. This function currently works for human, rat,
and mouse.
}
\details{
This function is designed to perform all annotation loading, parsing, and
sequence extractions for the reference of interest.

\code{load_annotations} uses the \code{ensembldb} package for handling annotations.
This allows for the use of \link[AnnotationFilter:AnnotationFilter]{AnnotationFilter::AnnotationFilter}s for
easily selecting transcripts of interest.

Most of the arguments to \code{load_annotations} are geared towards customizing
the output features and sequence extracted from the \link[ensembldb:EnsDb]{ensembldb::EnsDb}
object. Filtering parameters are inherited from
\code{\link[=build_annotation_filter]{build_annotation_filter()}}`.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
anno.db = load_annotations("hg38")
\dontshow{\}) # examplesIf}
}
