% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example DESeq2 data into the environment}
\usage{
load_example_data(example.type = c("sirna", "mirna"))
}
\arguments{
\item{example.type}{Name of the example to load. Options: sirna, mirna}
}
\value{
Loads either the Schlegel 2022 RNAseq data or miRDB into the environment.
}
\description{
Load example DESeq2 data into the environment
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
load_example_data()
\dontshow{\}) # examplesIf}
}
