% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_distances.R
\name{match_distances}
\alias{match_distances}
\title{Find inter match distances for each gene}
\usage{
match_distances(
  gr_list,
  feature.pos = c("fivePrimeEnd", "threePrimeEnd", "center")
)
}
\arguments{
\item{gr_list}{A GRangesList of matches from SeedMatchR}

\item{feature.pos}{Which position to use for central point. Options: "fivePrimeEnd", "threePrimeEnd", "center"}
}
\value{
A list of calculated distances for each match in the GRanges object grouped by gene name.
}
\description{
Find inter match distances for each gene
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
match_distances(gr)
\dontshow{\}) # examplesIf}
}
