% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_matches.R
\name{message_passing}
\alias{message_passing}
\title{message_passing}
\usage{
message_passing(graph, transcription_network, agg = "mean")
}
\arguments{
\item{graph}{An igraph object representing the graph.}

\item{transcription_network}{A data frame containing the transcription network with columns 'target' and 'edge_type'.}

\item{agg}{A character string specifying the aggregation method. Options are 'mean' or 'degree_importance'. Default is 'mean'.}
}
\value{
The updated graph with new vertex attributes 'feature_rep' and 'features_act'.
}
\description{
This function performs message passing on a graph based on a transcription network and an aggregation method.
}
