% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seeds.R
\name{plot_seeds}
\alias{plot_seeds}
\title{Plot the Guide Strand with different optional seeds}
\usage{
plot_seeds(guide.seq)
}
\arguments{
\item{guide.seq}{Guide a.k.a anti-sense sequence oriented 5' > 3'. Sequence must
be greater than 8 bp.}
}
\value{
A msaggplot of the guide sequence in addition to the available
seed sequences
}
\description{
Plot the Guide Strand with different optional seeds
}
\examples{
library(msa)

# Ttr siRNA sequence
guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

# generate seed plot
plotted.seeds = plot_seeds(guide.seq)
}
