% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{Plot matches as alignments against target sequence feature}
\usage{
plotAlignment(
  match.info,
  guide.seq,
  seqs,
  aln.name = "alignment",
  left.pad = 0,
  right.pad = 0,
  seq.start.pos = 2,
  seq.stop.pos = 8,
  output.type = c("data.frame", "ggplot")
)
}
\arguments{
\item{match.info}{A row from the \code{data.frame} of matches produced by \code{SeedMatchR}}

\item{guide.seq}{The full siRNA guide sequence}

\item{seqs}{The feature sequences used in scanning}

\item{aln.name}{The column name for the \code{data.frame} reporting alignments}

\item{left.pad}{The number of nucleotides to the left to show on the feature sequence beyond the length of the guide sequence.}

\item{right.pad}{The number of nucleotides to the right to show on the feature sequence beyond the length of the guide sequence.}

\item{seq.start.pos}{The seed definition start position}

\item{seq.stop.pos}{The seed definition stop position}

\item{output.type}{Either \code{data.frame} or \code{ggplot}}
}
\value{
A \code{data.frame} representing the input sequences and alignment between them.
}
\description{
Plot matches as alignments against target sequence feature
}
