% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wobble_bulges_pipeline.R
\name{run_wobble_bulges_pipeline}
\alias{run_wobble_bulges_pipeline}
\title{Running wobble and bulge inclusive seedmatchR pipeline}
\usage{
run_wobble_bulges_pipeline(
  guide.seq,
  res = res,
  features = features,
  bulges_lim = 4,
  nt = 3,
  silent = FALSE
)
}
\arguments{
\item{guide.seq}{A character string representing the guide sequence.}

\item{res}{A DESeq2 results data.frame.}

\item{features}{A DNAStringSet object with sequence information for features. The names of the sequences should be the transcript names.}

\item{bulges_lim}{An integer representing the number of target bulges allowed per sequence. Default is 4.}

\item{nt}{An integer representing the length from the start of siRNA where wobbles are not allowed. Default is 3.}

\item{silent}{A logical value to determine if the function should print progress. Default is FALSE.}
}
\value{
A data.frame with the overall matches for each seed.
}
\description{
This function runs a wobble and bulge inclusive SeedMatchR pipeline.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
result_df = run_wobble_bulges_pipeline(guide.seq = guide.seq, res = res, features = features)
\dontshow{\}) # examplesIf}
}
