% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedMatchReport.R
\name{SeedMatchReport}
\alias{SeedMatchReport}
\title{Generate statistics based on sequence matches and DE data}
\usage{
SeedMatchReport(
  res,
  seqs,
  guide.seq,
  group.name = "SeedMatchReport",
  indel.bool = TRUE,
  allow_wobbles = FALSE,
  log2fc.cut = -0.7655347,
  padj.cut = 0.05,
  seed.edit.distances = 0,
  fixed = T
)
}
\arguments{
\item{res}{A \code{data.frame} of \code{DESeq2} results}

\item{seqs}{A \code{Biostrings::DNAStringSet} of sequences to search}

\item{guide.seq}{Input RNA sequence as a character vector}

\item{group.name}{The group name to use for labeling rows}

\item{indel.bool}{Boolean in long format (TRUE, FALSE) for whether to include indels.}

\item{allow_wobbles}{Whether to allow G:U wobbles in the search.}

\item{log2fc.cut}{Log2(FC) threshold for considering a gene an off-target.}

\item{padj.cut}{Adjust p-value threshold}

\item{seed.edit.distances}{An integer value indicating allowed edit distance for seed searches.}

\item{fixed}{Whether IUPAC values need to match for ambiguous chracters.}
}
\value{
A list containing:
report: a \code{data.frame} of results
table: a \code{gt::gt} table of results
matches: A data.frame of matches
}
\description{
Generate statistics based on sequence matches and DE data
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
#' guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

anno.db = load_annotations("rnor7")

# Load test data
get_example_data("sirna")

sirna.data = load_example_data("sirna")

res <- sirna.data$Schlegel_2022_Ttr_D1_30mkg

# Filter DESeq2 results for SeedMatchR
res = filter_res(res, fdr.cutoff=1, fc.cutoff=0, rm.na.log2fc = TRUE)

SeedMatchReport(res, anno.db$seqs, anno.db$gtf, guide.seq, "si-Ttr")
\dontshow{\}) # examplesIf}
}
