% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_hist.R
\name{sequence_kmer_counts}
\alias{sequence_kmer_counts}
\title{Count the k-mer frequency in a set of sequences}
\usage{
sequence_kmer_counts(
  seqs,
  tx2gene,
  out.path = NULL,
  path.basename = NULL,
  width = 7,
  step = 1,
  simplify.as = "matrix"
)
}
\arguments{
\item{seqs}{Input \link[Biostrings:XStringSet-class]{Biostrings::DNAStringSet}}

\item{tx2gene}{Data frame mapping tx > gene with columsn tx_id + gene_id}

\item{out.path}{Output directory path}

\item{path.basename}{Output basename}

\item{width}{K-mer size}

\item{step}{step size}

\item{simplify.as}{output type as either matrix or list}
}
\value{
oligonucleotide frequencies
}
\description{
Count the k-mer frequency in a set of sequences
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
