% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmatchPattern2.R
\name{vmatchPattern2}
\alias{vmatchPattern2}
\title{A modified vmatchPattern function that supports indels and mismatches}
\usage{
vmatchPattern2(
  pattern,
  subject,
  max.mismatch = 0,
  min.mismatch = 0,
  with.indels = FALSE,
  fixed = TRUE,
  algorithm = "auto"
)
}
\arguments{
\item{pattern}{String to search for}

\item{subject}{DNAStringSet to query}

\item{max.mismatch}{Maximum number of mismatches allowed}

\item{min.mismatch}{Minimum number of mismatches allowed}

\item{with.indels}{Boolean for whether to include indels}

\item{fixed}{Boolean for whether to match character symbols}

\item{algorithm}{Which search algorithm to use. Defaults to "auto"}
}
\value{
A data frame of interval ranges per sequence
}
\description{
A modified vmatchPattern function that supports indels and mismatches
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
print("TBD")
\dontshow{\}) # examplesIf}
}
