% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_fc_ecdf.R
\name{deseq_fc_ecdf}
\alias{deseq_fc_ecdf}
\title{Plot the ECDF of log2(fold change) between groups}
\usage{
deseq_fc_ecdf(
  res,
  gene_lists,
  title = "ECDF",
  output_filename = NULL,
  palette = SeedMatchR.palette,
  factor_order = NULL,
  x_lims = c(-1, 1),
  stats_test = c("KS", "Wilcoxen", "Wass", "DTS"),
  alternative = c("greater", "less", "two.sided"),
  null_name = 1,
  target_name = 2,
  height = 5,
  width = 5,
  dpi = 320,
  theme = SeedMatchR.theme
)
}
\arguments{
\item{res}{The differential expression results data frame.}

\item{gene_lists}{A nested list of gene names. Example:
gene_lists = list("Background" = background_list, "Target" = target_list)}

\item{title}{The tile of the plot.}

\item{output_filename}{If the output filename is provided, then the plot is
saved.}

\item{palette}{Vector of colors to use for your curves.}

\item{factor_order}{Vector defining the order to use for the legends.}

\item{x_lims}{The x-axis range limits.}

\item{stats_test}{The statistical test to use. Options: KS, Wass, Wilcoxen, or
DTS.}

\item{alternative}{The alternative hypothesis to test. Options: greater,
less, two.sided.}

\item{null_name}{The name in the gene.list to use as the null for ECDF plots.}

\item{target_name}{The name in the gene.list to use as the target for ECDF.}

\item{height}{Plot height in inches.}

\item{width}{Plot width in inches.}

\item{dpi}{The DPI resolution for the figure.}

\item{theme}{A ggplot2 theme defining plot parameters.}
}
\value{
A ggplot object and ECDF statistics.
}
\description{
This functions will take differential expression results as a
\code{data.frame} and plot the ECDF for the input \code{gene_lists}.

The gene sets to plot should be provided as a named list of vectors.

Example:

\code{gene_lists = list("Background" = c("gene1", "gene2"), "Target" = c("gene2", "gene3"))}

The output will be saved to a PDF if an \code{output_filename} is provided.

Users can define the groups that are to be compared in the statistical test
using the \code{null_name} and \code{target_name} arguments. The names must be found
in \code{gene_lists}. The \code{factor_order} is used to order the groups in the
analysis.

This functions returns:
\itemize{
\item \verb{$plot}: The ECDF plot
\item \verb{$stats}: The results object
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(dplyr)

guide.seq = "UUAUAGAGCAAGAACACUGUUUU"

anno.db = load_annotations("rnor7")

# Load test data
get_example_data("sirna")

sirna.data = load_example_data("sirna")

res <- sirna.data$Schlegel_2022_Ttr_D1_30mkg

# Filter DESeq2 results for SeedMatchR
res = filter_res(res, fdr.cutoff=1, fc.cutoff=0, rm.na.log2fc = TRUE)

res = SeedMatchR(res = res, gtf = anno.db$gtf, seqs = anno.db$seqs,
sequence = guide.seq, seed.name = "mer7m8", tx.id.col= FALSE)

# Gene set 1
mer7m8.list = res$gene_id[res$mer7m8 >= 1]

# Gene set 2
background.list = res$gene_id[res$mer7m8 == 0]

ecdf.results = deseq_fc_ecdf(res,
list("Background" = background.list, "mer7m8" = mer7m8.list))
\dontshow{\}) # examplesIf}
}
