% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Download example DESeq2 data from GEO}
\usage{
get_example_data(example.type)
}
\arguments{
\item{example.type}{Name of the example to load. Options: sirna, mirna}
}
\description{
This function will download 3 files from GEO that can be used as toy datasets for SeedMatchR.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_example_data()
\dontshow{\}) # examplesIf}
}
