% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_annotations.R
\name{load_species_anno_db}
\alias{load_species_anno_db}
\title{Load species-specific annotations from \code{AnnotationHub}}
\usage{
load_species_anno_db(
  reference.name = c("hg38", "hg38-old", "mm39", "mm10", "rnor6", "rnor7", "rnor7.113"),
  filter = NULL
)
}
\arguments{
\item{reference.name}{Reference build name. Options: hg38, rnor6, rnor7,
mm39, mm10}

\item{filter}{An \link[AnnotationFilter:AnnotationFilterList]{AnnotationFilter::AnnotationFilterList} used to refine annotations.}
}
\value{
The function will return:
\itemize{
\item \verb{$gtf}: A \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges} object containing the GTF information
\item \verb{$dna}: A \link[rtracklayer:TwoBitFile-class]{rtracklayer::TwoBitFile} with reference DNA sequence
\item \verb{$txdb}: A \link[ensembldb:EnsDb]{ensembldb::EnsDb} object with transcriptome information
}
}
\description{
Use \code{\link[AnnotationHub:AnnotationHub-class]{AnnotationHub::AnnotationHub()}} to load species-specific
\link[rtracklayer:TwoBitFile-class]{rtracklayer::TwoBitFile} DNA sequence and \link[ensembldb:EnsDb]{ensembldb::EnsDb}
object. This function currently works for human, rat, and mouse.

The specific versions and builds used for each reference are further defined
in \code{articles(SeedMatchR_Annotation_Databases)}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
anno.db = SeedMatchR::load_species_anno_db("hg38")
\dontshow{\}) # examplesIf}
}
